backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic[email]
passlib
bcrypt
PyJWT
google-auth
google-auth-oauthlib
requests
python-multipart
python-dotenv
razorpay
stripe
reportlab
sendgrid
pytest
httpx
